[feature] extract UserResumeService interface
